⚡ Updated draw method

+++ android/src/main/java/com/adrbrpa/capacitorheatmap/data/models/Point.kt
@@ -1,3 +1,3 @@
 package com.adrbrpa.capacitorheatmap.data.models
 
-data class Point(val x: Double, val y: Double, val thickness: Double)
+data class Point(val x: Float, val y: Float, val thickness: Float)
